@@ -1,6 +1,6 @@
 name: better_player
 description: Advanced video player based on video_player and Chewie. It's solves many typical use cases and it's easy to run.
-version: 0.0.83
+version: 0.0.84
 # Disabled because of warning from analyzer
 # authors:
 #  - Jakub Homlala <jhomlala@gmail.com>
@@ -8,17 +8,17 @@ homepage: https://github.com/jhomlala/betterplayer
 documentation: https://jhomlala.github.io/betterplayer/
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.16.0 <3.0.0'
   flutter: ">=2.2.3"
 
 dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.5
-  wakelock: ^0.6.1+2
+  wakelock: ^0.6.2
   meta: ^1.7.0
   flutter_widget_from_html_core: ^0.8.5+3
-  visibility_detector: ^0.3.3
+  visibility_detector: ^0.4.0+2
   path_provider: ^2.0.10
   collection: ^1.16.0
   xml: ^6.1.0
