@@ -34,9 +34,9 @@ class MockMethodChannel {
   void _createEventChannel(int id) {
     final MethodChannel eventChannel =
         MethodChannel("better_player_channel/videoEvents$id");
-    TestDefaultBinaryMessengerBinding.instance!.defaultBinaryMessenger
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
         .setMockMethodCallHandler(eventChannel, (MethodCall methodCall) async {
-      TestDefaultBinaryMessengerBinding.instance!.defaultBinaryMessenger
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
           .handlePlatformMessage(
               "better_player_channel/videoEvents$id",
               const StandardMethodCodec()
